# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  invoice.InvoiceItems:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      discount:
        type: integer
      invoice_type:
        allOf:
        - $ref: '#/definitions/invoice.InvoiceType'
        enum:
        - per_hour
        - per_unit
      item_id:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    required:
    - amount
    - description
    - invoice_type
    - quantity
    - unit_price
    type: object
  invoice.InvoiceType:
    enum:
    - per_unit
    - per_hour
    type: string
    x-enum-varnames:
    - PerUnit
    - PerHour
...
  recurring.CreateRecurringDTO:
    properties:
      country:
        type: string
      interval:
        allOf:
        - $ref: '#/definitions/recurring.RecurringInterval'
        enum:
        - weekly
        - biweekly
        - monthly
        - quarterly
        - yearly
      invoice_items:
        items:
          $ref: '#/definitions/invoice.InvoiceItems'
        minItems: 1
        type: array
      note:
        type: string
      payer_email:
        type: string
      payer_name:
        type: string
      service_fee:
        minimum: 0
        type: number
      start_date:
        type: string
      template_id:
        type: string
      title:
        minLength: 3
        type: string
    required:
    - country
    - interval
    - invoice_items
    - payer_email
    - payer_name
    - start_date
    - template_id
    - title
    type: object
  recurring.RecurringInterval:
    enum:
    - weekly
    - biweekly
    - monthly
    - quarterly
    - yearly
    type: string
    x-enum-varnames:
    - IntervalWeekly
    - IntervalBiweekly
    - IntervalMonthly
    - IntervalQuarterly
    - IntervalYearly
  recurring.UpdateRecurringDTO:
    properties:
      interval:
        allOf:
        - $ref: '#/definitions/recurring.RecurringInterval'
        enum:
        - weekly
        - biweekly
        - monthly
        - quarterly
        - yearly
      is_active:
        type: boolean
      start_date:
        type: string
    type: object
  user.CompleteProfileRequestDto:
    properties:
      business_name:
        maxLength: 50
        type: string
      country:
        maxLength: 25
        type: string
      first_name:
        maxLength: 15
        minLength: 2
        type: string
      last_name:
        maxLength: 15
        minLength: 2
        type: string
      phone_number:
        maxLength: 15
        type: string
    required:
    - first_name
    - last_name
    type: object
  user.UpdateProfileDto:
    properties:
      business_name:
        maxLength: 50
        type: string
      country:
        maxLength: 25
        type: string
      first_name:
        maxLength: 15
        minLength: 2
        type: string
      last_name:
        maxLength: 15
        minLength: 2
        type: string
      phone_number:
        maxLength: 15
        type: string
    type: object
  utils.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:4000
info:
  contact:
    email: davidhero125@gmail.com
    name: API Support
  description: Invoice management and payment platform API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Stellance API
  version: "1.0"
paths:
...
  /recurring:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: List recurring invoice schedules
      tags:
      - recurring
    post:
      consumes:
      - application/json
      parameters:
      - description: Recurring invoice config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/recurring.CreateRecurringDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create recurring invoice schedule
      tags:
      - recurring
  /recurring/{id}:
    delete:
      parameters:
      - description: Recurring ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete recurring schedule
      tags:
      - recurring
    get:
      parameters:
      - description: Recurring ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get recurring schedule by ID
      tags:
      - recurring
    patch:
      consumes:
      - application/json
      parameters:
      - description: Recurring ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/recurring.UpdateRecurringDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update recurring schedule
      tags:
      - recurring
  /transactions:
    get:
      parameters:
      - description: Filter by user ID (admin only)
        in: query
        name: user_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - transactions
  /transactions/{id}:
    delete:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete a transaction
      tags:
      - transactions
  /transactions/id/{id}:
    get:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - transactions
  /transactions/inflow:
    get:
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get cash flow over a date range
      tags:
      - transactions
  /transactions/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get transaction overview stats
      tags:
      - transactions
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Profile data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.CompleteProfileRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Complete user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateProfileDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - user
  /users/{id}:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get user profile by ID
      tags:
      - user
  /users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get current authenticated user
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"